Remove password fields when logged in with Github
If a user has used Github authentication to create their account,
they are not able to update their password as one does not exist.
This resolves the issue by removing the Change Password form, and
also the Password input on the Edit Profile page.

One possible improvement that could be made in future would be
additional checks to ensure the user is still authenticated with
Github, though I've spent all day on this already so I'll leave
that to somebody else to add in ;)