multi allowed vendors for copyright watchers
I don't know if Symfony has multiple, but `flarum-lang` (for Flarum Languages) receives so many updates with the amount of languages being added lately. This solves the issue by marking certain vendors as allowed.

I haven't tested this by the way, but I might try doing that if you'd be up for this.